rag-backend/requirements.txt
# RAG Backend Requirements

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
python-multipart>=0.0.5

# Document Processing
Pypdf2 # For PDF processing
# pip3 install PyPDF2
langchain>=0.0.300

# Knowledge Graph Dependencies
spacy>=3.0.0
networkx>=2.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# LLM Dependencies
google.generativeai>=0.3.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Utilities
requests>=2.26.0
tqdm>=4.65.0
